fix : ts 파일에서 tsx 파일로 변경
